+++ NEWS.md
@@ -3,6 +3,11 @@
 ## Added
 
 * site_name and run_name are printed to console in `growR_run_loop`.
+* PhenologicalAutocut
+
+## Changed
+
+* autocut algorithm is now outsourced to its own R6class.
 
 # growR 1.3.0